\docType{package}
\name{additivityTests-package}
\alias{additivityTests}
\alias{additivityTests-package}
\title{Additivity tests in the two way ANOVA with single sub-class numbers.}
\description{
  Testing for interaction in the two way ANOVA with single
  sub-class numbers.
}
\author{
  Petr Simecek <simecek@gmail.com>
}
\keyword{package}

