\name{Boik}
\alias{Boik}
\docType{data}
\title{Multi-headed Machine Data}
\description{
 Performance of a multiple-headed machine used to fill bottles. Weights for six heads on five occasions were recorded. 
}
\usage{data(Boik)}
\source{
  Robert J. Boik: A comparison of three invariant tests of additivity in two-way classifications with no replications, Computational Statistics \& Data Analysis, 1993.
}
\examples{
data(Boik)
Boik
}
\keyword{datasets}
