\docType{data}
\name{Boik}
\alias{Boik}
\title{Multi-headed Machine Data}
\source{
  Robert J. Boik: A comparison of three invariant tests of
  additivity in two-way classifications with no
  replications, Computational Statistics \& Data Analysis,
  1993.
}
\description{
  Performance of a multiple-headed machine used to fill
  bottles. Weights for six heads on five occasions were
  recorded.
}
\keyword{datasets}

